Update dependencies, use better folder structure and keep this mother fucking project alive. There is still hope
