+++ .gitignore
@@ -44,5 +44,3 @@ yarn-error.log*
 /*.ipr
 /*.iws
 
-# vscode
-.vscode
