+++ .prettierrc.json
@@ -1,5 +1,5 @@
 {
-    "useTabs": false,
-    "tabWidth": 4,
-    "singleQuote": false
+  "semi": true,
+  "singleQuote": false,
+  "trailingComma": "none"
 }
