--- components/Footer/Footer.tsx
@@ -1,9 +1,0 @@
-const Footer = () => {
-    return (
-        <div>
-            <p>Footer</p>
-        </div>
-    );
-}
-
-export default Footer;
