+++ package.json
@@ -1,31 +1,32 @@
 {
-    "name": "motimaa.net",
-    "private": true,
-    "scripts": {
-        "dev": "next dev",
-        "build": "next build",
-        "start": "next start",
-        "lint": "next lint"
-    },
-    "dependencies": {
-        "bootstrap": "^5.1.3",
-        "next": "12.1.0",
-        "next-themes": "^0.1.1",
-        "react": "17.0.2",
-        "react-bootstrap": "^2.2.1",
-        "react-dom": "17.0.2",
-        "sass": "^1.49.9"
-    },
-    "devDependencies": {
-        "@types/node": "17.0.23",
-        "@types/react": "17.0.42",
-        "eslint": "8.11.0",
-        "eslint-config-next": "12.1.0",
-        "prettier": "^2.6.0",
-        "stylelint": "^14.6.0",
-        "stylelint-config-prettier-scss": "^0.0.1",
-        "stylelint-config-standard-scss": "^3.0.0",
-        "stylelint-order": "^5.0.0",
-        "typescript": "4.6.2"
-    }
+  "name": "motimaa.net",
+  "private": true,
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint",
+    "pretty": "prettier --write \"./**/*.{ts,tsx,json}\""
+  },
+  "dependencies": {
+    "bootstrap": "^5.1.3",
+    "next": "^12.1.5",
+    "next-themes": "^0.1.1",
+    "react": "^18.0.0",
+    "react-bootstrap": "^2.3.0",
+    "react-dom": "^18.0.0",
+    "sass": "^1.50.1"
+  },
+  "devDependencies": {
+    "@types/node": "^17.0.25",
+    "@types/react": "^18.0.6",
+    "eslint": "^8.14.0",
+    "eslint-config-next": "^12.1.5",
+    "prettier": "^2.6.0",
+    "stylelint": "^14.6.0",
+    "stylelint-config-prettier-scss": "^0.0.1",
+    "stylelint-config-standard-scss": "^3.0.0",
+    "stylelint-order": "^5.0.0",
+    "typescript": "^4.6.3"
+  }
 }
