+++ src/client/components/Footer/Footer.tsx
@@ -1,0 +1,9 @@
+const Footer = () => {
+  return (
+    <div>
+      <p>Footer</p>
+    </div>
+  );
+};
+
+export default Footer;
