+++ src/pages/_document.tsx
@@ -1,5 +1,5 @@
 // pages/_document.js
-import {Head, Html, Main, NextScript} from "next/document";
+import { Head, Html, Main, NextScript } from "next/document";
 
 export default function Document() {
   return (
