+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
-import type {NextPage} from "next";
-import styles from "../styles/Home.module.css";
-import Navbar from "../components/Navbar/Navbar";
-import {Col, Container, Row} from "react-bootstrap";
+import Navbar from "client/components/Navbar/Navbar";
+import styles from "client/styles/Home.module.css";
+import type { NextPage } from "next";
+import { Col, Container, Row } from "react-bootstrap";
 
 const Home: NextPage = () => {
   return (
